レスの記事が存在しない時の動作の調整。

+++ petitnote/index.php
@@ -27,8 +27,8 @@ require_once(__DIR__.'/noticemail.inc');
 //テンプレート
 $skindir='template/'.$skindir;
 
-$petit_ver='v0.18.6';
-$petit_lot='lot.220516';
+$petit_ver='v0.18.8';
+$petit_lot='lot.220521';
 
 if(!isset($functions_ver)||$functions_ver<20220515){
 	return error($en?'Please update functions.php to the latest version.':'functions.phpを最新版に更新してください。');
@@ -509,7 +509,11 @@ function post(){
 	$newline='';
 	if($resto){//レスの時はスレッド別ログに追記
 		$r_line = "$resto\t$sub\t$name\t$verified\t$com\t$url\t$imgfile\t$w\t$h\t$thumbnail\t$painttime\t$img_md5\t$tool\t$pchext\t$time\t$time\t$host\t$userid\t$hash\tres\n";
-		file_put_contents(LOG_DIR.$resto.'.log',$r_line,FILE_APPEND | LOCK_EX);
+		if(is_file("LOG_DIR{$resto}.log")){
+			file_put_contents(LOG_DIR.$resto.'.log',$r_line,FILE_APPEND | LOCK_EX);
+		}else{
+			return error($en? 'The article does not exist.':'記事がありません。');
+		}
 		chmod(LOG_DIR.$resto.'.log',0600);
 		if(!$sage){
 			foreach($alllog_arr as $i =>$val){
@@ -1915,6 +1919,9 @@ function view($page=0){
 				$out[$oya][]=$_res;
 			}	
 		fclose($rp);
+		if(empty($out[$oya])||$out[$oya][0]['oya']!=='oya'){
+			unset($out[$oya]);
+		}
 	}
 
 	//管理者判定処理
